Bug 4430 - More helpful error message when the initial condition for a species is higher than the max molecules per molecular type limit in the solver options panel (for NFSim). Set limit to 200000.
svn path=/vcell/trunk/VCell/; revision=18342

+++ src/org/vcell/solver/nfsim/NFsimXMLWriter.java
@@ -832,7 +832,7 @@ public class NFsimXMLWriter {
 				double value = evaluateConstant(initialCount.getCount(),simulationSymbolTable);
 				Integer maxMoleculesPerType = simulationSymbolTable.getSimulation().getSolverTaskDescription().getNFSimSimulationOptions().getMaxMoleculesPerType();
 				if (maxMoleculesPerType == null){
-					maxMoleculesPerType = new Integer(100000);
+					maxMoleculesPerType = new Integer(200000);
 				}
 				if (maxMoleculesPerType.doubleValue() < value){
 					String eMessage = "The Initial count for Species '" + seedSpecies.getName() + "' is " + BigDecimal.valueOf(value).toBigInteger();
